requirements.txt
fastapi
uvicorn
pandas
python-multipart
python-dotenv
google-generativeai